cleanup sqlite3_wrapper

+++ src/net/manager.cpp
@@ -113,7 +113,7 @@ void game_manager::tick() {
         if (l.lifetime <= ticks{0}) {
             broadcast_message<"lobby_removed">(l.lobby_id);
             m_lobbies.erase(*it);
-            tracking::track_lobby_count(static_cast<int>(m_lobbies.size()));
+            tracking::track_lobby_count(m_lobbies.size());
             it = m_lobby_order.erase(it);
             continue;
         }
@@ -193,7 +193,7 @@ void game_manager::handle_message(utils::tag<"lobby_make">, game_user &user, con
 
     id_type lobby_id = ++m_lobby_count;
     auto &l = m_lobby_order.emplace_back(m_lobbies.try_emplace(lobby_id, value, lobby_id).first)->second;
-    tracking::track_lobby_count(static_cast<int>(m_lobbies.size()));
+    tracking::track_lobby_count(m_lobbies.size());
 
     int user_id = l.add_user(user).user_id;
 
@@ -310,7 +310,7 @@ void game_manager::kick_user_from_lobby(game_user &user) {
 void game_manager::on_connect(client_handle client) {
     logging::info("{}: Connected", get_client_ip(client));
     m_clients.emplace(client, client);
-    tracking::track_client_count(static_cast<int>(m_clients.size()));
+    tracking::track_client_count(m_clients.size());
 }
 
 void game_manager::on_disconnect(client_handle client) {
@@ -323,7 +323,7 @@ void game_manager::on_disconnect(client_handle client) {
             }
         }
         m_clients.erase(it);
-        tracking::track_client_count(static_cast<int>(m_clients.size()));
+        tracking::track_client_count(m_clients.size());
     }
 }
